requirements.txt
feedparser~=6.0.2
newspaper3k~=0.2.8
nltk~=3.6.2
requests~=2.25.1
Pillow~=8.2.0
boto3~=1.17.85